tests/wpt/web-platform-tests/old-tests/submission/Opera/script_scheduling/106-noimport.html

<!DOCTYPE html>
<html><head>
        <title> scheduler: stylesheets blocking scripts</title>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="testlib/testlib.js"></script>
        <!-- this stylesheet blocks scripts -->
        <link rel="stylesheet" href="css/background.css?pipe=trickle(d2)">
</head>
<body>
        <div id="log">FAILED (This TC requires JavaScript enabled)</div>
        <div id="test">Test</div>

        <script>
          test(function() {
            assert_equals(getComputedStyle(document.getElementById("test")).position,
                          "fixed");
          });
        </script>
</body></html>
